Lab/ms_code_lab_day_1_4/pyproject.toml
[project]
name = "summerschool-2025"
version = "0.1.0"
description = "Summerschool 2025 labs with FastAPI backend and Streamlit UI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.112",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "httpx>=0.27",
  "streamlit>=1.31",
  "ruff>=0.5",
  "openai>=1.106.1",
  "python-dotenv>=1.1.1",
  "markitdown[all]>=0.1.3",
  "qdrant-client>=1.15.1",
  "langchain>=0.3.27",
  "langchain-community>=0.3.29",
  "pypdf>=6.0.0",
  "langgraph>=0.6.7",
  "langchain-qdrant>=0.2.1",
  "sentence-transformers>=5.1.0",
  "langchain-huggingface>=0.3.1",
  "langchain-openrouter>=0.0.1",
]

[tool.uv]
package = false
